tratamiento de errores de None

+++ model/model_peliculas.py
@@ -9,24 +9,39 @@ def ver_peliculas():
 #Filtrar los registros de una tabla por id (solo que sea exactamente igual a)
 def encontrar_pelicula(id_pelicula):
     pelicula = peliculas.query.filter(peliculas.idPelicula == id_pelicula).first()
-    print(pelicula)
+    if pelicula is not None:
+        print(pelicula)
+    else:
+        print("La pelicula con id = "+ str(id_pelicula) + " no existe")
     
 #Actualizar la columna nombre de un registro
 def cambiar_nombre_peli_por_id(id_pelicula, nombre_nuevo):
     pelicula = peliculas.query.filter(peliculas.idPelicula == id_pelicula).first()
-    pelicula.nombre = nombre_nuevo
-    db.session.commit()
+    if pelicula is not None:
+        pelicula.nombre = nombre_nuevo
+        db.session.commit()
+    else:
+        print("La pelicula con id = "+ str(id_pelicula) + " no existe")
+
     
 def cambiar_nombre_peli_por_nombre(nombre, nombre_nuevo):
     pelicula = peliculas.query.filter(peliculas.nombre == nombre).first()
-    pelicula.nombre = nombre_nuevo
-    db.session.commit()
+    if pelicula is not None:
+        pelicula.nombre = nombre_nuevo
+        db.session.commit()
+    else:
+        print("La pelicula con nombre = \""+ nombre + "\" no existe")
+
     
 #Eliminar un registro por id
 def eliminar_pelicula(id_pelicula):
     pelicula = peliculas.query.filter(peliculas.idPelicula == id_pelicula).first()
-    db.session.delete(pelicula)
-    db.session.commit()
+    if pelicula is not None:
+        db.session.delete(pelicula)
+        db.session.commit()
+    else:
+        print("La pelicula con id = "+ str(id_pelicula) + " no existe")
+
     
 #Eliminar todos los registros
 def eliminar_todas_las_peliculas():
