+++ model/model_rentar.py
@@ -9,19 +9,31 @@ def ver_rentas():
 #Filtrar los registros de una tabla por id (solo que sea exactamente igual a)
 def encontrar_renta(id_renta):
     renta = rentar.query.filter(rentar.idRentar == id_renta).first()
-    print(renta)
+    if renta is not None:
+        print(renta)
+    else:
+        print("La renta con id = "+ str(id_renta) + " no existe")
+
     
 #Modificar la fecha de la renta
 def cambiar_fecha_renta(id_renta, fecha_nueva):
     renta = rentar.query.filter(rentar.idRentar == id_renta).first()
-    renta.fecha_renta = fecha_nueva
-    db.session.commit()
-    
+    if renta is not None:
+        renta.fecha_renta = fecha_nueva
+        db.session.commit()
+    else:
+        print("La renta con id = "+ str(id_renta) + " no existe")
+
+        
 #Eliminar un registro por id
 def eliminar_renta(id_renta):
     renta = rentar.query.filter(rentar.idRentar == id_renta).first()
-    db.session.delete(renta)
-    db.session.commit() 
+    if renta is not None:
+        db.session.delete(renta)
+        db.session.commit() 
+    else:
+        print("La renta con id = "+ str(id_renta) + " no existe")
+
 
 #Eliminar todos los registros
 def eliminar_todas_las_rentas():
